Rename observers module
