+++ tardis/plugins/sqliteregistry.py
